requirements.txt
# Core Django
Django==6.0
social-auth-app-django==5.6.0

# Database
PyMySQL==1.1.1

# Wiki integration
pywikibot==9.5.0
mwclient==0.11.0
mwparserfromhell==0.6.6

# REST API
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Security checks
bandit==1.8.0
safety==3.2.11

# Type checking
mypy==1.13.0
django-stubs==5.1.1
djangorestframework-stubs==3.15.1
types-requests==2.32.0.20241016
